refactor/ajustando valor de quantidade de estoque para double e retor…

+++ src/main/java/com/hortifruti/hortifrutiapi/dto/produto/ProdutoEstoqueDTO.java
@@ -3,7 +3,7 @@ package com.hortifruti.hortifrutiapi.dto.produto;
 public interface ProdutoEstoqueDTO {
     String getNomeProduto();
     String getUnidadeMedida();
-    Integer getQuantidade();
+    Double getQuantidade();
     String getNomeSede();
 }
